Appium is like Selenium – but for mobile apps and games “. You’ve probably read/heard this several times before, but in fact, Appium is much more than that. Appium is also suitable for mobile web testing when real devices and real browsers are used in testing.
 “Appium is like Selenium – but for mobile apps and games“. You’ve
probably read/heard this several times before, but in fact, Appium is
much more than that. Appium is also suitable for mobile web testing when
real devices and real browsers are used in testing. Emulators won’t do
the trick as e.g. Chrome cannot be even installed on standard Android
QEMU-based emulator. Some emulators may work (with Chrome), but why to
bother – those are still emulators and no single end-user uses emulators
to run your app, game or web products. With the same effort (and cost)
you can build your test automation with Appium using the real stuff.

In a nutshell, Appium is a mobile test automation framework (with a
tool) that works for all: native, hybrid and mobile web apps for iOS and
Android. Appium is a great choice for test automation framework as it
can be used for all these different app/web types. Basically, Appium
derives its roots from Selenium and it uses JSONWireProtocol internally
to interact with iOS and Android apps using Selenium’s WebDriver.

In its architecture, Appium is an HTTP server written in Node.js that
creates and handles multiple WebDriver sessions. Appium starts tests on
the device and listens for commands from the main Appium server. It is
basically the same as the Selenium server that gets HTTP requests from
Selenium client libraries.